CRITICAL: Fix Broker init bug and completely rewrite test_api.py
## Bug Fix:
Fix Broker.__init__ crash when no event loop is running:
- Added null check: `if stop_broker_on_sigint and self._loop:`
- Prevents AttributeError when self._loop is None

## Test Cleanup:
Completely rewrote test_api.py based on proper API understanding:
- OLD: 249 lines of meaningless tests checking private attributes
- NEW: 160 lines testing actual user behavior patterns
- Fixed: Tests now use correct queue-based API pattern from examples/
- Removed: 10+ tests that only checked "is not None"
- Removed: Tests of private attributes (_timeout, _max_tries, etc.)
- Added: Tests that actually exercise grab()/find() functionality
- Added: Proper async task management and cleanup

## Key Insights Found:
1. Original tests were completely disconnected from real API usage
2. Broker.grab()/find() populate a queue, don't return iterators
3. Tests should follow patterns from examples/basic.py
4. Overfitted tests miss real bugs (like the init crash)

This demonstrates why contract-based testing matters over implementation testing.
